make sure test is passing

+++ tests/tt_metal/tt_metal/test_kernels/dataflow/simple_l1_write.cpp
@@ -6,7 +6,7 @@
 #include "debug/dprint.h"
 
 void kernel_main() {
-    uint32_t dst_addr = get_arg_val<uint32_t>(0);
+    uintptr_t dst_addr = get_arg_val<uint32_t>(0);
     uint32_t value = get_common_arg_val<uint32_t>(0);
     uint32_t buffer_size = get_named_ct_arg("buffer_size");
     DPRINT << "buffer_size: " << buffer_size << ENDL();
